pythonProject/requirements.txt
Flask==3.0.3
pymongo==4.10.1
python-dotenv~=1.0.1
pytest==7.2.2